rename to filterOpsAndEstimateGas
